add cart

+++ src/main/java/vn/ute/smartphoneshop/entity/CartDetailEntity.java
@@ -24,5 +24,5 @@ public class CartDetailEntity {
     private int quantity;
 
     @Column(name = "cartPrice", nullable = false)
-    private double cartPrice;
+    private Long cartPrice;
 }
